+++ R/lsoa_to_msoa_lookup.R
@@ -9,44 +9,27 @@
 #' @keywords internal
 #' @return a data frame (a lookup table)
 #'
-lsoa_to_msoa_lookup <- function(
-                                df,
+lsoa_to_msoa_lookup <- function(df,
                                 nmw = TRUE,
                                 keep = FALSE) {
 
 
-  create_msoa_lookup <- function(keep_nmw = TRUE) {
+  msoa_cols <- 1:5
+  # don't include Welsh language names (codes ending "nmw")
+  if (!nmw) msoa_cols <- c(1:2, 4)
 
-    hocl_msoa_names <- function(nmw = TRUE) {
-      msoa_cols <- 1:5
-      # don't include Welsh language names (codes ending "nmw")
-      if (!nmw) msoa_cols <- c(1:2, 4)
+  hocl_msoa_names <- hocl_msoa_names %>%
+    dplyr::select(dplyr::all_of(msoa_cols))
 
-      # The House of Commons Library's friendly MSOA names
-      # https://visual.parliament.uk/msoanames
-      paste0(
-        "https://visual.parliament.uk/msoanames/static/",
-        "MSOA-Names-Latest.csv") %>%
-        readr::read_csv() %>%
-        dplyr::select(dplyr::all_of(msoa_cols))
-    }
-
-    # see data-raw/datasets.R for source
-    lsoa11cdnm %>%
-      dplyr::mutate(
-        msoa11nm = stringr::str_remove(lsoa11nm, "[A-Z]{1}$")
-      ) %>%
-      dplyr::left_join(hocl_msoa_names(nmw = keep_nmw)) %>%
-      dplyr::relocate(msoa11nm, .after = msoa11cd)
-  }
-
-
-  # get msoa lookup (above)
-  msoa_lookup <- create_msoa_lookup(keep_nmw = nmw)
+  msoa_lookup <- lsoa11cdnm %>%
+    dplyr::mutate(
+      msoa11nm = stringr::str_remove(lsoa11nm, "[A-Z]{1}$")
+    ) %>%
+    dplyr::left_join(hocl_msoa_names) %>%
+    dplyr::relocate(msoa11nm, .after = msoa11cd)
 
 
   df_out <- df %>%
-    # dplyr::select(dplyr::any_of(c("oa11cd", "lsoa11cd", "lsoa11nm"))) %>%
     dplyr::select(dplyr::starts_with(c("oa", "lsoa"))) %>%
     dplyr::distinct() %>%
     dplyr::left_join(msoa_lookup) %>%