+++ R/upper_tier_region_ctry_lookup.R
@@ -1,7 +1,9 @@
-#' (data) A lookup of upper-tier Local Authorities (EW) to their Region (if any) and Country
+#' A lookup of UTLAs (EW) to their Region (E) and Country
 #'
-#' @format A data frame with 174 rows and 6 variables:
+#' @format A data frame with 331 rows and 8 variables:
 #' \describe{
+#'   \item{ltla21cd}{LTLA code}
+#'   \item{ltla21cd}{LTLA name}
 #'   \item{utla21cd}{UTLA code}
 #'   \item{utla21nm}{UTLA name}
 #'   \item{rgn21cd}{Region code}