+++ tests/testthat/test-build_api_query.R
@@ -3,8 +3,8 @@ test_that(
     expect_identical(
       build_api_query(
         ref = 4,
-        within_level = "cauth20nm",
-        within = "Greater Manchester",
+        where_level = "cauth20nm",
+        where = "Greater Manchester",
         fields = c("lad20cd", "lad20nm", "cauth20cd", "cauth20nm")
       ),
 
@@ -21,8 +21,8 @@ test_that(
     expect_identical(
       build_api_query(
         ref = 10,
-        within_level = "msoa11cd",
-        within = c("E02006666", "E02006667", "E02006668"),
+        where_level = "msoa11cd",
+        where = c("E02006666", "E02006667", "E02006668"),
         fields = c("msoa11cd", "msoa11nm")
       ),
 
@@ -39,13 +39,13 @@ test_that(
     expect_identical(
       build_api_query(
         ref = 12,
-        within_level = "lad20nm",
-        within = c("Cheltenham", "Gloucester", "Stroud", "Cotswold", "Tewkesbury", "Forest of Dean"),
+        where_level = "lad20nm",
+        where = c("Cheltenham", "Gloucester", "Stroud", "Cotswold", "Tewkesbury", "Forest of Dean"),
         fields = c("lad20cd", "lad20nm", "BNG_E", "BNG_N"),
         sr = 7405
       ),
 
-      "https://services1.arcgis.com/ESMARspQHYMw9BZ9/arcgis/rest/services/Local_Authority_Districts_December_2020_UK_BFC/FeatureServer/0/query?where=%20(LAD20NM%20=%20'CHELTENHAM'%20OR%20LAD20NM%20=%20'GLOUCESTER'%20OR%20LAD20NM%20=%20'STROUD'%20OR%20LAD20NM%20=%20'COTSWOLD'%20OR%20LAD20NM%20=%20'TEWKESBURY'%20OR%20LAD20NM%20=%20'FOREST%20OF%20DEAN')%20&outFields=LAD20CD,LAD20NM,BNG_E,BNG_N&outSR=7405&resultType=none&returnDistinctValues=true&f=json"
+      "https://services1.arcgis.com/ESMARspQHYMw9BZ9/arcgis/rest/services/LAD_MAY_2021_UK_BGC/FeatureServer/0/query?where=%20(LAD20NM%20=%20'CHELTENHAM'%20OR%20LAD20NM%20=%20'GLOUCESTER'%20OR%20LAD20NM%20=%20'STROUD'%20OR%20LAD20NM%20=%20'COTSWOLD'%20OR%20LAD20NM%20=%20'TEWKESBURY'%20OR%20LAD20NM%20=%20'FOREST%20OF%20DEAN')%20&outFields=LAD20CD,LAD20NM,BNG_E,BNG_N&outSR=7405&resultType=none&returnDistinctValues=true&f=json"
     )
   }
 )